requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
python-dotenv==1.0.1
SQLAlchemy==2.0.35
pandas==2.2.2
plotly==5.24.1
dash==2.18.2
azure-eventhub==5.12.2
click==8.1.7
pytest==8.3.2
